Move Game class to shared objects, more lobby stuff
