+++ CommSubSystem/CommSubSystem.csproj
@@ -40,9 +40,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Conversation\Conversation.cs" />
-    <Compile Include="Conversation\ConversationFactory.cs" />
-    <Compile Include="Conversation\CreateGameConv.cs" />
     <Compile Include="TCPClient.cs" />
     <Compile Include="Commands\Command.cs" />
     <Compile Include="Commands\CommandFactory.cs" />
@@ -56,7 +53,9 @@
     <Compile Include="Commands\RespCommand.cs" />
     <Compile Include="Receive\ReceivingFactory.cs" />
     <Compile Include="Receive\ReceiveInvoker.cs" />
-    <Compile Include="Conversation\GameListConv.cs" />
+    <Compile Include="ConversationClass\Conversation.cs" />
+    <Compile Include="ConversationClass\ConversationFactory.cs" />
+    <Compile Include="ConversationClass\CreateGameConv.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ConsoleApp1\Utils.csproj">
