+++ SharedObjects/SharedObjects.csproj
@@ -56,6 +56,7 @@
     <Compile Include="ProcessType.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PublicEndPoint.cs" />
+    <Compile Include="Game.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
