+++ SushiGo.userprefs
@@ -1,14 +1,16 @@
 ﻿<Properties StartupConfiguration="{EEB75A52-775C-47FC-9C2B-21210F0DFB1E}|Default">
-  <MonoDevelop.Ide.Workbench ActiveDocument="LobbyApp/Lobby.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="SharedObjects/Game.cs">
     <Files>
-      <File FileName="LobbyApp/LobbyForm.cs" Line="69" Column="1" />
-      <File FileName="LobbyApp/Lobby.cs" Line="54" Column="77" />
-      <File FileName="SharedObjects/Player.cs" Line="18" Column="27" />
-      <File FileName="UserApp/ClientForm.cs" Line="72" Column="45" />
-      <File FileName="CommSubSystem/Conversation/CreateGameConv.cs" Line="15" Column="13" />
-      <File FileName="CommSubSystem/Conversation/GameListConv.cs" Line="11" Column="38" />
-      <File FileName="Messages/JoinGame.cs" Line="7" Column="9" />
+      <File FileName="LobbyApp/LobbyForm.cs" Line="67" Column="32" />
+      <File FileName="LobbyApp/Lobby.cs" Line="89" Column="6" />
+      <File FileName="SharedObjects/Player.cs" Line="17" Column="26" />
+      <File FileName="UserApp/ClientForm.cs" Line="83" Column="13" />
       <File FileName="Messages/LobbyHeartbeat.cs" Line="8" Column="6" />
+      <File FileName="CommSubSystem/ConversationClass/CreateGameConv.cs" Line="40" Column="41" />
+      <File FileName="CommSubSystem/Receive/ReceivingFactory.cs" Line="15" Column="18" />
+      <File FileName="Messages/CreateGame.cs" Line="12" Column="44" />
+      <File FileName="Messages/SendGameList.cs" Line="7" Column="1" />
+      <File FileName="SharedObjects/Game.cs" Line="15" Column="30" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
